logic mockup tự chọn với số 2

+++ config/livewire.php
@@ -3,7 +3,7 @@
 return [
     'temporary_file_upload' => [
         'disk' => null,
-        'rules' => ['required', 'file', 'max:'.env('LIVEWIRE_UPLOAD_MAX_KB', 204800)],
+        'rules' => ['required', 'file', 'max:'.(int) env('LIVEWIRE_UPLOAD_MAX_KB', 204800)],
         'directory' => null,
         'middleware' => null,
         'preview_mimes' => [
@@ -11,7 +11,7 @@ return [
             'mov', 'avi', 'wmv', 'mp3', 'm4a',
             'jpg', 'jpeg', 'mpga', 'webp', 'wma',
         ],
-        'max_upload_time' => env('LIVEWIRE_UPLOAD_MAX_MINUTES', 15),
+        'max_upload_time' => (int) env('LIVEWIRE_UPLOAD_MAX_MINUTES', 15),
         'cleanup' => true,
     ],
 ];
